Moved previous app.py down to CDTADPQ.web
